style: Interface and implementation separation
